backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Snikls <macks.chigrin@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.5"
psycopg2-binary = "^2.9.10"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.4.0"
drf-yasg = "^1.21.8"
drf-spectacular = "^0.28.0"
requests = "^2.32.3"
black = "^24.10.0"
isort = "^6.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
